# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/account/orders": {
      "get": {
        "operationId": "listOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderSummaryListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Pedidos do cliente logado",
        "tags": [
          "account"
        ]
      }
    },
    "/api/v1/account/addresses": {
      "get": {
        "operationId": "listAddresses",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "account"
        ]
      },
      "post": {
        "operationId": "createAddress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "account"
        ]
      }
    },
    "/api/v1/account/addresses/{id}": {
      "put": {
        "operationId": "updateAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "account"
        ]
      },
      "delete": {
        "operationId": "deleteAddress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "account"
        ]
      }
    },
...
    "/api/v1/admin/orders": {
      "get": {
        "operationId": "listOrders",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 80,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending_payment",
                "paid",
                "processing",
                "shipped",
                "delivered",
                "cancelled",
                "refunded"
              ]
            }
          },
          {
            "name": "paymentMethod",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pix",
                "boleto",
                "credit_card"
              ]
            }
          },
          {
            "name": "paymentStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "confirmed",
                "received",
                "overdue",
                "refunded",
                "cancelled"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedOrdersDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Lista pedidos com filtros por status, pagamento, período e busca",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vellor API",
    "description": "API da loja Vellor (relógios de luxo e perfumes de nicho). Autenticação por cookie de sessão httpOnly; erros no formato RFC 7807 (application/problem+json).",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "vellor_sid"
      }
    },
    "schemas": {
...
      "OkDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "OrderSummaryListDto": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "number": {
              "type": "string"
            },
            "status": {
              "type": "string",
              "enum": [
                "pending_payment",
                "paid",
                "processing",
                "shipped",
                "delivered",
                "cancelled",
                "refunded"
              ]
            },
            "statusLabel": {
              "type": "string"
            },
            "totalCents": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991
            },
            "itemCount": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991
            },
            "paymentMethod": {
              "type": "string",
              "enum": [
                "pix",
                "boleto",
                "credit_card"
              ]
            },
            "paymentStatus": {
              "type": "string",
              "enum": [
                "pending",
                "confirmed",
                "received",
                "overdue",
                "refunded",
                "cancelled"
              ]
            },
            "customerName": {
              "type": "string"
            },
            "customerEmail": {
              "type": "string"
            },
            "createdAt": {
              "type": "string"
            }
          },
          "required": [
            "id",
            "number",
            "status",
            "statusLabel",
            "totalCents",
            "itemCount",
            "paymentMethod",
            "paymentStatus",
            "customerName",
            "customerEmail",
            "createdAt"
          ]
        }
      },
      "AddressListDto": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "recipientName": {
              "type": "string",
              "minLength": 3,
              "maxLength": 120
            },
            "cep": {
              "type": "string"
            },
            "street": {
              "type": "string",
              "minLength": 2,
              "maxLength": 160
            },
            "number": {
              "type": "string",
              "minLength": 1,
              "maxLength": 20
            },
            "complement": {
              "type": "string",
              "maxLength": 80,
              "nullable": true
            },
            "district": {
              "type": "string",
              "minLength": 2,
              "maxLength": 80
            },
            "city": {
              "type": "string",
              "minLength": 2,
              "maxLength": 80
            },
            "state": {
              "type": "string",
              "enum": [
                "AC",
                "AL",
                "AP",
                "AM",
                "BA",
                "CE",
                "DF",
                "ES",
                "GO",
                "MA",
                "MT",
                "MS",
                "MG",
                "PA",
                "PB",
                "PR",
                "PE",
                "PI",
                "RJ",
                "RN",
                "RS",
                "RO",
                "RR",
                "SC",
                "SP",
                "SE",
                "TO"
              ]
            },
            "reference": {
              "type": "string",
              "maxLength": 160,
              "nullable": true
            },
            "id": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "label": {
              "type": [
                "string",
                "null"
              ]
            },
            "isDefault": {
              "type": "boolean"
            }
          },
          "required": [
            "cep",
            "street",
            "number",
            "district",
            "city",
            "state",
            "id",
            "label",
            "isDefault"
          ]
        }
      },
      "SaveAddressDto": {
        "type": "object",
        "properties": {
          "recipientName": {
            "type": "string",
            "minLength": 3,
            "maxLength": 120
          },
          "cep": {
            "type": "string"
          },
          "street": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160
          },
          "number": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20
          },
          "complement": {
            "type": "string",
            "maxLength": 80,
            "nullable": true
          },
          "district": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          },
          "city": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          },
          "state": {
            "type": "string",
            "enum": [
              "AC",
              "AL",
              "AP",
              "AM",
              "BA",
              "CE",
              "DF",
              "ES",
              "GO",
              "MA",
              "MT",
              "MS",
              "MG",
              "PA",
              "PB",
              "PR",
              "PE",
              "PI",
              "RJ",
              "RN",
              "RS",
              "RO",
              "RR",
              "SC",
              "SP",
              "SE",
              "TO"
            ]
          },
          "reference": {
            "type": "string",
            "maxLength": 160,
            "nullable": true
          },
          "label": {
            "type": "string",
            "maxLength": 40,
            "nullable": true
          },
          "isDefault": {
            "type": "boolean"
          }
        },
        "required": [
          "cep",
          "street",
          "number",
          "district",
          "city",
          "state"
        ]
      },
      "AddressDto": {
        "type": "object",
        "properties": {
          "recipientName": {
            "type": "string",
            "minLength": 3,
            "maxLength": 120
          },
          "cep": {
            "type": "string"
          },
          "street": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160
          },
          "number": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20
          },
          "complement": {
            "type": "string",
            "maxLength": 80,
            "nullable": true
          },
          "district": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          },
          "city": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          },
          "state": {
            "type": "string",
            "enum": [
              "AC",
              "AL",
              "AP",
              "AM",
              "BA",
              "CE",
              "DF",
              "ES",
              "GO",
              "MA",
              "MT",
              "MS",
              "MG",
              "PA",
              "PB",
              "PR",
              "PE",
              "PI",
              "RJ",
              "RN",
              "RS",
              "RO",
              "RR",
              "SC",
              "SP",
              "SE",
              "TO"
            ]
          },
          "reference": {
            "type": "string",
            "maxLength": 160,
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "label": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isDefault": {
            "type": "boolean"
          }
        },
        "required": [
          "cep",
          "street",
          "number",
          "district",
          "city",
          "state",
          "id",
          "label",
          "isDefault"
        ]
      },
...
      "PaginatedOrdersDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "number": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "pending_payment",
                    "paid",
                    "processing",
                    "shipped",
                    "delivered",
                    "cancelled",
                    "refunded"
                  ]
                },
                "statusLabel": {
                  "type": "string"
                },
                "totalCents": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "itemCount": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "paymentMethod": {
                  "type": "string",
                  "enum": [
                    "pix",
                    "boleto",
                    "credit_card"
                  ]
                },
                "paymentStatus": {
                  "type": "string",
                  "enum": [
                    "pending",
                    "confirmed",
                    "received",
                    "overdue",
                    "refunded",
                    "cancelled"
                  ]
                },
                "customerName": {
                  "type": "string"
                },
                "customerEmail": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "number",
                "status",
                "statusLabel",
                "totalCents",
                "itemCount",
                "paymentMethod",
                "paymentStatus",
                "customerName",
                "customerEmail",
                "createdAt"
              ]
            }
          },
          "page": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "pageSize": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "total": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "totalPages": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
    }
  }
}